requirements.txt
langchain==0.3.26
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.9
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-openai===0.3.24
langchain-community==0.3.26

-e .